+++ packages/react-wrap-with/src/wrapWith.extract.noContainer.spec.tsx
@@ -1,0 +1,23 @@
+/** @jest-environment jsdom */
+/// <reference types="@types/jest" />
+
+import { render } from '@testing-library/react';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import wrapWith from './wrapWith';
+
+const Hello = ({ className }) => <h1 className={className}>Hello, World!</h1>;
+
+Hello.propTypes = { className: PropTypes.string };
+
+test('extract props without a container', () => {
+  // GIVEN: Wrapping <Hello> with false and extract "className" prop.
+  const BlinkingHello = wrapWith(false, {}, ['className'])(Hello);
+
+  // WHEN: Render.
+  const result = render(<BlinkingHello className="blink" />);
+
+  // THEN: It should produce HTML without class name because it is already extracted.
+  expect(result.container.innerHTML).toMatchInlineSnapshot(`"<h1>Hello, World!</h1>"`);
+});