+++ packages/react-wrap-with/src/wrapWith.refObject.classComponent.spec.tsx
@@ -2,7 +2,7 @@
 /// <reference types="@types/jest" />
 
 import { render } from '@testing-library/react';
-import React, { Component, forwardRef, useEffect, useRef } from 'react';
+import React, { Component, useEffect, useRef } from 'react';
 
 import wrapWith from './wrapWith';